vm: fix gas refund bug on evm message error
